pyproject.toml
[tool.poetry]
name = "skypro-diplom"
version = "0.1.0"
description = ""
authors = ["Oleg <oleg.a.matyash@gmail.com>"]
readme = "README.md"
packages = [{include = "skypro_diplom"}]

[tool.poetry.dependencies]
python = "^3.9"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.3"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"



[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tools.isort]
py_version = 310
profile = 'django'
line_length = 120
multi_line_output = 2
know_first_party = ["todolist", "core"]

[tool.flake8]
max-line-length = 120
per-file-ignores = """
    */migrations/*: E501,
"""
